% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage21S223-package.R
\docType{package}
\name{testpackage21S223-package}
\alias{testpackage21S223}
\alias{testpackage21S223-package}
\title{testpackage21S223: Demo creating a package for 413/613-001}
\description{
Not very much right now.
}
\author{
\strong{Maintainer}: ChiYun Liu \email{cl0471a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
